pyproject.toml
[project]
name = "dotfiles-test-harness"
version = "0.1.0"
description = "Governanca, testes e automacao do worktree de validacao dos dotfiles."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.3",
]

[dependency-groups]
dev = [
  "pymarkdownlnt>=0.9.35",
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "ruff>=0.15.2",
  "ty>=0.0.18",
  "yamllint>=1.37.1",
]

[tool.uv]
package = false
cache-dir = ".cache/uv"
required-version = ">=0.10.8,<0.11"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=scripts --cov-report=term-missing"
testpaths = ["tests/python"]
cache_dir = ".cache/pytest"
norecursedirs = [".cache", ".git", ".venv", "archive"]

[tool.ruff]
line-length = 100
target-version = "py311"
cache-dir = ".cache/ruff"
respect-gitignore = true
extend-exclude = [".cache", ".venv", "archive"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B904", "E501", "UP017"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.src]
include = ["scripts", "tests/python"]
exclude = [
  ".venv",
  ".cache",
  "__pycache__",
  "archive",
]

[tool.ty.terminal]
error-on-warning = false

[tool.ty.rules]
unresolved-import = "error"

[tool.coverage.run]
data_file = ".cache/coverage/.coverage"
